server/requirements.txt
fastapi
SQLAlchemy
pytesseract
python-dotenv
opencv-python-headless
selenium
requests
webdriver_manager
celery[redis]
numpy
pydantic
uvicorn
starlette